+++ Packages/com.github.kurotu.vrc-quest-tools/Editor/VRCQuestTools.cs
@@ -27,7 +27,7 @@ namespace KRT.VRCQuestTools
         /// <summary>
         /// VRCQuestTools version (semver).
         /// </summary>
-        public const string Version = "2.3.3";
+        public const string Version = "2.3.4";
 
         /// <summary>
         /// GitHub repository (username/reponame).